Create check.py
A sales executive needs to simplify an assigned sales task. The task is to sell the items given to her in a bag, where each item has an ID number.
It is easier to sell items with the same ID numbers. The sales executive also decides to remove some items from the bag. The largest number of items that she can remove is also known to her.
Find the minimum number of different IDs the final bag can contain after removing the allowed number of items.